move source files into assembler/interpreter, Update + improve READMEs
Improve the READMEs according to chatgpt3.5 suggestions.
