+++ flipjump/interpretter/io_devices/FixedIO.py
@@ -1,19 +1,13 @@
-from sys import stdin, stdout
-
-from flipjump.io_devices.IODevice import IODevice
-from flipjump.inner_classes.exceptions import IOReadOnEOF, IncompleteOutput
-from flipjump.utils.constants import io_bytes_encoding
+from flipjump.interpretter.io_devices.IODevice import IODevice
+from flipjump.utils.exceptions import IOReadOnEOF, IncompleteOutput
 
 
-class StandardIO(IODevice):
+class FixedIO(IODevice):
     """
-    read from stdin, write to stdout
+    read from fixed input, don't output (with get_output functionality)
     """
-    def __init__(self, output_verbose: bool):
-        """
-        @param output_verbose: if true print program's output
-        """
-        self.output_verbose = output_verbose
+    def __init__(self, _input: bytes):
+        self.remaining_input = _input
         self._output = b''
 
         self.current_input_byte = 0
@@ -24,11 +18,11 @@ class StandardIO(IODevice):
 
     def read_bit(self) -> bool:
         if 0 == self.bits_to_read_in_input_byte:
-            read_bytes = stdin.read(1).encode(encoding=io_bytes_encoding)
-            if 0 == len(read_bytes):
-                raise IOReadOnEOF("Read an empty input on standard IO (EOF)")
+            if not self.remaining_input:
+                raise IOReadOnEOF("Read an empty input on fixed IO (EOF)")
 
-            self.current_input_byte = read_bytes[0]
+            self.current_input_byte = self.remaining_input[0]
+            self.remaining_input = self.remaining_input[1:]
             self.bits_to_read_in_input_byte = 8
 
         bit = (self.current_input_byte & 1) == 1
@@ -41,15 +35,15 @@ class StandardIO(IODevice):
         self.bits_to_write_in_output_byte += 1
 
         if 8 == self.bits_to_write_in_output_byte:
-            curr_output: bytes = self.current_output_byte.to_bytes(1, 'little')
-            if self.output_verbose:
-                stdout.write(curr_output.decode(encoding=io_bytes_encoding))
-                stdout.flush()
-            self._output += curr_output
+            self._output += self.current_output_byte.to_bytes(1, 'little')
             self.current_output_byte = 0
             self.bits_to_write_in_output_byte = 0
 
     def get_output(self, *, allow_incomplete_output=False) -> bytes:
+        """
+        @raise IncompleteOutput when the number of outputted bits can't be divided by 8
+        @return: full output until now
+        """
         if not allow_incomplete_output and 0 != self.bits_to_write_in_output_byte:
             raise IncompleteOutput("tries to get output when an unaligned number of bits was outputted "
                                    "(doesn't divide 8)")
