+++ programs/calc.fj
@@ -1,3 +1,5 @@
+// This program is a calculator supports the [+-*/%] operations between two numbers (decimal / hexadecimal).
+
 stl.startup
