+++ programs/print_tests/print_dec.fj
@@ -1,3 +1,5 @@
+// This program prints the numbers v1,v2,v3 in decimal form, each in its own line.
+
 stl.startup
 
 print_int v1
