+++ tests/README.md
@@ -6,15 +6,15 @@ run `pytest` to run the fast tests.
 
 Run with `--compile` / `--run` for testing only the compilation / the run.
 
-Add a combination of `--fast`, `--medium`, `--slow`, `--hexlib` to run tests of different types.<br>
-Use `--regulr` to run all the tests of stable parts.<br>
-Use `--all` to run all the tests.<br>
+Add a combination of `--fast`, `--medium`, `--slow`, `--hexlib` to run tests of different types.  
+Use `--regulr` to run all the tests of stable parts.  
+**====> Use `--all` to run all the tests. <====**  
 The default (no type flags) means `--fast`.
 
 ![Running Pytest with --regular](../res/pytest.gif)
 
-You can run the tests parallel with `-n auto` (using [xdist](https://github.com/pytest-dev/pytest-xdist)).<br>
-note that this option is only allowed while using exactly one of `--compile` / `--run`.<br>
+You can run the tests parallel with `-n auto` (using [xdist](https://github.com/pytest-dev/pytest-xdist)).  
+note that this option is only allowed while using exactly one of `--compile` / `--run`.  
 You can execute the `test_parallel` / `test_parallel.bat` to run parallel compile, and afterwords parallel run, with the given flags.
 
 ![Running the test_parallel script with --regular](../res/test_parallel.gif)
@@ -52,8 +52,10 @@ You can get the entire "macro-stack" of the last executed addresses to get a bet
 More information about the labels can be found [here](../flipjump/README.md#generated-label-names).
 
 ### Filter tests by their name 
-Using the next tests together will take the union of the relevant tests.
- * `--name n1 n2` only run tests with these names.
+
+You can filter the running tests. No filter means that all the tests in the chosen test type (--regular, --all, etc.) will run.  
+Using the next filters together will take the union of the relevant tests.
+ * `--name n1 n2` only run tests with these exact names.
  * `--contains n1 n2` only run tests containing one of these names.
  * `--startswith n1 n2` only run tests starting with one of these names.
  * `--endswith n1 n2` only run tests ending with one of these names.
@@ -62,10 +64,14 @@ Using the next tests together will take the union of the relevant tests.
 
 The .csv files in this directory specify which tests to run (and what parameters to use). 
 
-To support more .csv files, update [conf.json](conf.json).<br>
+A test is a .fj program that gets compiled into a .fjm file, which gets run with a predefined input, and checks if it finished successfully with the desired output.
+
+To support more .csv files, update [conf.json](conf.json).  
 The python test itself can be found on [test_fj.py](test_fj.py) (and [conftest.py](conftest.py)).
 
-To add a new test, first choose the relevant csv file.<br>
+To add a new test, create your .fj programs, and you predefined input and expected output.
+
+Then choose the relevant csv file.  
 The rule of thumb (for the sum of compile+run times, in seconds):
 
 | fast         | medium       | slow |
@@ -82,6 +88,8 @@ Then add a new line to the relevant compile-csv and run-csv files, according to 
 |--------------|-------------------------------------------------------------|------------------------------|--------------|---------|-------|---------|--------------------------|
 | example_test | path/to/example_1.fj &#124; ... &#124; path/to/example_n.fj | path/to/compiled/example.fjm | 64           | 1       | 0     | True    | True                     |
 
+Every line in this file will cause the compilation of the specified .fj files into the .fjm file, with the given compilation parameters.
+
 Note that you can specify a single file, or a '|' separated list of files in the .fj paths cell. In case of a list, they will be compiled in the inserted order.
 
 ### Run CSVs format:
@@ -91,6 +99,8 @@ Note that you can specify a single file, or a '|' separated list of files in the
 | test name    | .fjm path                    | input file path           | output file path            | is input a binary file | is output a binary file |
 |--------------|------------------------------|---------------------------|-----------------------------|------------------------|-------------------------|
 | example_test | path/to/compiled/example.fjm | path/to/inputs/example.in | path/to/outputs/example.out | False                  | False                   |
+
+Every line in this file will cause the specified .fjm to run with the given input, and its output will be compared to the given expected output.
 
 Note that you can also emit specifying a file in the input/output cell, and leave it empty. In that case an empty input/output will be used.
